+++ src/graphics/ui/text.py
@@ -12,7 +12,7 @@ class Text(Drawable):
             text: str,
             position: tuple[float, float],
             sprite_sheet: SpriteSheet,
-            text_size: int,
+            text_size: float,
             anchor: str = "center",
             ) -> None:
         super().__init__(position, sprite_sheet, anchor)
@@ -28,10 +28,13 @@ class Text(Drawable):
 
         self._surface = pygame.Surface((self.__width, self.__height))
         # import random
-        # self._surface.fill((random.randrange(0, 255), random.randrange(0, 255), random.randrange(0, 255)))
+        # self._surface.fill(
+        #     (random.randrange(0, 255),
+        #      random.randrange(0, 255),
+        #      random.randrange(0, 255)
+        #      )
+        #     )
         self.__font.render(self._surface, (0, 0), self.__text)
 
     def getSize(self) -> tuple[float, float]:
         return (self.__width, self.__height)
-
-
